dont throw an exception during parseResponse(), or the buffer wont get eaten up.
git-svn-id: https://gearbox.svn.sf.net/svnroot/gearbox/gearbox/trunk@333 4b7f92db-2845-0410-95f1-b1f669fef64b

+++ src/gbxsickacfr/driver.cpp
@@ -309,7 +309,7 @@ Driver::desiredAngularResolution()
     assert( angleIncrementInHundredthDegrees == ANGULAR_RESOLUTION_1_0_DEG ||
             angleIncrementInHundredthDegrees == ANGULAR_RESOLUTION_0_5_DEG ||
             angleIncrementInHundredthDegrees == ANGULAR_RESOLUTION_0_25_DEG );
-    return angleIncrementInHundredthDegrees;
+    return (uint16_t)angleIncrementInHundredthDegrees;
 }
 
 bool
